+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `Ruzzy.c_trace_branch` to `Ruzzy.trace` to simplify interface
+
 ## [0.6.0] - 2024-02-13
 
 ### Added